requirements.txt
Django==2.0.7
django-registration-redux==2.4
pytz==2018.5
